+++ RFEM6_Adapter/CRUD/Delete/_IDelete.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2022, the respective contributors. All rights reserved.
  *
@@ -20,28 +20,31 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using BH.oM.Base;
-using BH.oM.Base.Attributes;
+using BH.oM.Adapter;
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace BH.oM.Adapters.SoftwareName
+namespace BH.Adapter.RFEM6
 {
-    [Description("Object description in here. Will appear in the UI tooltip.")]
-    public class ExampleObject : BHoMObject
+    public partial class RFEM6Adapter : BHoMAdapter
     {
-        // // See examples in the BHoM repo and the wiki to see how we define types.
-        // // Generally, all properties should be public and have public getter and setter.
-        // // BHoM Objects should have orthogonal properties and no behaviour (no methods), as in C# Records (or Python Dataclasses).
-        // // No constructor should be specified. If a specific instantiaton method is needed, we make it as an "Engine/Create" method.
-        // // Objects created with this convention will automatically appear as UI components (e.g. Grasshopper component).
+        // Basic Delete method that deletes objects depending on their Type and Id. 
+        // It gets called by the Push or by the Remove Adapter Actions.
+        // Its implementation is facultative (not needed for a simple export/import scenario). 
+        // Toolkits need to implement (override) this only to get the full CRUD to work.
+        protected override int IDelete(Type type, IEnumerable<object> ids, ActionConfig actionConfig = null)
+        {
+            //Insert code here to enable deletion of specific types of objects with specific ids
+            BH.Engine.Base.Compute.RecordError($"Delete for objects of type {type.Name} is not implemented in {(this as dynamic).GetType().Name}.");
+            return 0;
+        }
 
-        [Description("Property description in here.")]
-        public string SomeStringProperty { get; set; }
+        // There are more virtual Delete methods you might want to override and implement.
+        // Check the base BHoM_Adapter solution and the wiki for more info.
 
-        [Description("Property description in here.")]
-        public int SomeNumberProperty { get; set; }
+        /***************************************************/
     }
 }
