+++ RFEM6_Toolkit.sln
@@ -1,13 +1,13 @@
-﻿
+
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.32901.82
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SoftwareName_Adapter", "SoftwareName_Adapter\SoftwareName_Adapter.csproj", "{F84BEFC2-C702-4A86-9626-C7FA48785606}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RFEM6_Adapter", "RFEM6_Adapter\RFEM6_Adapter.csproj", "{F84BEFC2-C702-4A86-9626-C7FA48785606}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SoftwareName_Engine", "SoftwareName_Engine\SoftwareName_Engine.csproj", "{C68B4D77-E46D-41C9-A72C-F29984D8AB4A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RFEM6_Engine", "RFEM6_Engine\RFEM6_Engine.csproj", "{C68B4D77-E46D-41C9-A72C-F29984D8AB4A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SoftwareName_oM", "SoftwareName_oM\SoftwareName_oM.csproj", "{FA23407E-1C08-4183-A02E-A57FA9D09BCA}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RFEM6_oM", "RFEM6_oM\RFEM6_oM.csproj", "{FA23407E-1C08-4183-A02E-A57FA9D09BCA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
